Move synapse workspace tests to their own directory
