+++ tests/smoke_test/test_placeholder.py
