+++ tests/unit_test/test_azure.py
