+++ tests/validation_test/test_appeals_has.py
@@ -1,21 +1,21 @@
 import pytest
-import workspace.testing.util.pipelineutils as pipelineutils
-import testing.util.constants as constants
+import tests.util.pipelineutils as pipelineutils
+import tests.util.constants as constants
 import warnings
 
-def test_pins_lpa_curated(credential_name, azure_credential, synapse_endpoint: str):
+def test_appeal_appeals_has(credential_name, azure_credential, synapse_endpoint: str):
 
     warnings.filterwarnings("ignore", category=DeprecationWarning) 
 
     # run the testing notebook
-    notebookname: str = "py_unit_tests_pins_lpa_curated"
+    notebookname: str = "py_unit_tests_has_appeals"
     
     notebook_raw_params = {
         "notebook": notebookname,
         "parameters": {
             "entity_name": {
                "type": "String",
-               "value": "pins-lpa",
+               "value": "appeal-has",
             },
             "std_db_name": {
                "type": "String",
@@ -31,16 +31,16 @@ def test_pins_lpa_curated(credential_name, azure_credential, synapse_endpoint: s
             },
             "std_table_name": {
                "type": "String",
-               "value": "pins_lpa",
+               "value": "sb_appeal_has",
             },
             "hrm_table_name": {
                "type": "String",
-               "value": "pins_lpa",
+               "value": "sb_appeal_has",
             },
             "curated_table_name": {
                "type": "String",
-               "value": "pins_lpa",
-            },
+               "value": "appeal_has",
+            }
         }
     }
     notebook_raw_params.update(constants.SPARK_POOL_CONFIG)
@@ -50,7 +50,7 @@ def test_pins_lpa_curated(credential_name, azure_credential, synapse_endpoint: s
     print("Notebook response *" +str(exitMessage) +"*")
     assert notebook_run_result == constants.NOTEBOOK_SUCCESS_STATUS 
     assert exitMessage == constants.NOTEBOOK_EXIT_CODE_SUCCESS
-    print("test_pins_lpa_curated Completed")
+    print("test_appeal_event Completed")
 
 @pytest.fixture(autouse=True)
 def run_before_and_after_tests():
