+++ tests/validation_test/test_nsip_project.py
@@ -1,20 +1,21 @@
 import pytest
-import testing.util.constants as constants
-import workspace.testing.util.pipelineutils as pipelineutils
+import tests.util.pipelineutils as pipelineutils
+import tests.util.constants as constants
 import warnings
 
-def test_appeal_document_notebook(credential_name, azure_credential, synapse_endpoint: str):
+def test_nsip_project_notebook(credential_name, azure_credential, synapse_endpoint: str):
 
     warnings.filterwarnings("ignore", category=DeprecationWarning) 
 
     # run the testing notebook
-    notebookname: str = "py_unit_tests_appeal_document"
+    notebookname: str = "py_unit_tests_nsip_project"
+    
     notebook_raw_params = {
         "notebook": notebookname,
         "parameters": {
             "entity_name": {
                "type": "String",
-               "value": "appeal-document",
+               "value": "nsip-project",
             },
             "std_db_name": {
                "type": "String",
@@ -30,19 +31,35 @@ def test_appeal_document_notebook(credential_name, azure_credential, synapse_end
             },
             "std_table_name": {
                "type": "String",
-               "value": "sb_appeal_document",
+               "value": "sb_nsip_project",
             },
             "hrm_table_name": {
                "type": "String",
-               "value": "sb_appeal_document",
+               "value": "sb_nsip_project",
             },
             "hrm_table_final": {
                "type": "String",
-               "value": "appeals_document_metadata",
+               "value": "nsip_project",
             },
             "curated_table_name": {
                "type": "String",
-               "value": "appeal_document",
+               "value": "nsip_project",
+            },
+            "curated_db_migration_name": {
+               "type": "String",
+               "value": "odw_curated_migration_db",
+            },
+            "curated_table_migration_name": {
+               "type": "String",
+               "value": "nsip_project",
+            },
+            "primary_key": {
+               "type": "String",
+               "value": "caseId",
+            },
+            "migration_primary_key": {
+               "type": "String",
+               "value": "caseReference",
             }
         }
     }
@@ -53,7 +70,7 @@ def test_appeal_document_notebook(credential_name, azure_credential, synapse_end
     print("Notebook response *" +str(exitMessage) +"*")
     assert notebook_run_result == constants.NOTEBOOK_SUCCESS_STATUS 
     assert exitMessage == constants.NOTEBOOK_EXIT_CODE_SUCCESS
-    print("test_appeal_document Completed")
+    print("test_nsip_project Completed")
 
 @pytest.fixture(autouse=True)
 def run_before_and_after_tests():
