+++ tests/validation_test/test_pins_lpa_curated.py
@@ -1,21 +1,21 @@
 import pytest
-import workspace.testing.util.pipelineutils as pipelineutils
-import testing.util.constants as constants
+import tests.util.pipelineutils as pipelineutils
+import tests.util.constants as constants
 import warnings
 
-def test_nsip_project_notebook(credential_name, azure_credential, synapse_endpoint: str):
+def test_pins_lpa_curated(credential_name, azure_credential, synapse_endpoint: str):
 
     warnings.filterwarnings("ignore", category=DeprecationWarning) 
 
     # run the testing notebook
-    notebookname: str = "py_unit_tests_nsip_project"
+    notebookname: str = "py_unit_tests_pins_lpa_curated"
     
     notebook_raw_params = {
         "notebook": notebookname,
         "parameters": {
             "entity_name": {
                "type": "String",
-               "value": "nsip-project",
+               "value": "pins-lpa",
             },
             "std_db_name": {
                "type": "String",
@@ -31,36 +31,16 @@ def test_nsip_project_notebook(credential_name, azure_credential, synapse_endpoi
             },
             "std_table_name": {
                "type": "String",
-               "value": "sb_nsip_project",
+               "value": "pins_lpa",
             },
             "hrm_table_name": {
                "type": "String",
-               "value": "sb_nsip_project",
-            },
-            "hrm_table_final": {
-               "type": "String",
-               "value": "nsip_project",
+               "value": "pins_lpa",
             },
             "curated_table_name": {
                "type": "String",
-               "value": "nsip_project",
+               "value": "pins_lpa",
             },
-            "curated_db_migration_name": {
-               "type": "String",
-               "value": "odw_curated_migration_db",
-            },
-            "curated_table_migration_name": {
-               "type": "String",
-               "value": "nsip_project",
-            },
-            "primary_key": {
-               "type": "String",
-               "value": "caseId",
-            },
-            "migration_primary_key": {
-               "type": "String",
-               "value": "caseReference",
-            }
         }
     }
     notebook_raw_params.update(constants.SPARK_POOL_CONFIG)
@@ -70,7 +50,7 @@ def test_nsip_project_notebook(credential_name, azure_credential, synapse_endpoi
     print("Notebook response *" +str(exitMessage) +"*")
     assert notebook_run_result == constants.NOTEBOOK_SUCCESS_STATUS 
     assert exitMessage == constants.NOTEBOOK_EXIT_CODE_SUCCESS
-    print("test_nsip_project Completed")
+    print("test_pins_lpa_curated Completed")
 
 @pytest.fixture(autouse=True)
 def run_before_and_after_tests():
