+++ tests/validation_test/test_relevant_representation.py
@@ -1,21 +1,21 @@
 import pytest
-import workspace.testing.util.pipelineutils as pipelineutils
-import testing.util.constants as constants
+import tests.util.pipelineutils as pipelineutils
+import tests.util.constants as constants
 import warnings
 
-def test_appeal_service_user(credential_name, azure_credential, synapse_endpoint: str):
+def test_relevant_representation_notebook(credential_name, azure_credential, synapse_endpoint: str):
 
     warnings.filterwarnings("ignore", category=DeprecationWarning) 
 
     # run the testing notebook
-    notebookname: str = "py_unit_tests_service_user"
+    notebookname: str = "py_unit_tests_relevant_representation"
     
     notebook_raw_params = {
         "notebook": notebookname,
         "parameters": {
             "entity_name": {
                "type": "String",
-               "value": "service-user",
+               "value": "nsip-representation",
             },
             "std_db_name": {
                "type": "String",
@@ -29,29 +29,33 @@ def test_appeal_service_user(credential_name, azure_credential, synapse_endpoint
                "type": "String",
                "value": "odw_curated_db",
             },
-             "curated_migration_db_name": {
+            "curated_migration_db_name": {
                "type": "String",
                "value": "odw_curated_migration_db",
             },
             "std_table_name": {
                "type": "String",
-               "value": "sb_service_user",
+               "value": "sb_nsip_representation",
             },
             "hrm_table_name": {
                "type": "String",
-               "value": "sb_service_user",
+               "value": "sb_nsip_representation",
             },
             "hrm_table_final": {
                "type": "String",
-               "value": "service_user",
+               "value": "nsip_representation",
             },
-            "service_user_appeal_table": {
+            "curated_table_name": {
                "type": "String",
-               "value": "appeal_service_user",
+               "value": "nsip_representation",
             },
-            "service_user_migration_table": {
-                "type": "String",
-                "value": "service_user"
+            "std_hzn_table_name": {
+               "type": "String",
+               "value": "horizon_nsip_relevant_representation",
+            },
+            "curated_migration_table_name": {
+               "type": "String",
+               "value": "nsip_representation",
             }
         }
     }
@@ -62,7 +66,7 @@ def test_appeal_service_user(credential_name, azure_credential, synapse_endpoint
     print("Notebook response *" +str(exitMessage) +"*")
     assert notebook_run_result == constants.NOTEBOOK_SUCCESS_STATUS 
     assert exitMessage == constants.NOTEBOOK_EXIT_CODE_SUCCESS
-    print("test_service_user Completed")
+    print("test_relevant_representation Completed")
 
 @pytest.fixture(autouse=True)
 def run_before_and_after_tests():
